agarwal_recovery_solution

Wrapper to function theis_solution (The recovery test is analyzed using the Theis solution)

agarwal_recovery_solution(ptest, a, t0, t)

Arguments

ptest
A pumping_test object

a
Slope of the straight line fitted to the drawdown data using the Cooper-Jacob approach

t0
Intercept of the straight line fitted to the drawdown data using the Cooper-Jacob approach

t
Numeric vector with time

Value

A numeric vector with the calculate drawdown

References

Agarwal, R. (1980). A New Method To Account For Producing Time Effects When Drawdown Type Curves Are Used To Analyze Pressure Buildup And Other Test Data. SPE Paper 9289.

See also

Examples

data(agarwal_recovery)
pump_history <- 14400
agarwal_corr <- recovery_equivalent_time(agarwal_recovery$t, agarwal_recovery$s,
                                pump_history)
ptest.agarwal <- pumping_test("Well1", Q = 0.02893519, r = 60, t = agarwal_corr$ta,
s = agarwal_corr$sa)
sol0 <- agarwal_recovery_solution_initial(ptest.agarwal)
scal <- agarwal_recovery_solution(ptest.agarwal, sol0$a, sol0$t, ptest.agarwal$t)
print(scal)

#>  [1] 0.2180857 0.3193969 0.3818021 0.4621039 0.5154323 0.5719063 0.6355445
#>  [8] 0.6800058 0.7411371 0.7830082 0.8391526 0.8764278 0.9035921 0.9412595
#> [15] 0.9665594