confint_jackniffe

Function to calculate the

confint_jackniffe(ptest, level = c(0.025, 0.975), slug = F)

Arguments

ptest
A pumping_test object

level
A numeric vector with the significance levels of the confidence intervals.

slug
A logical flag indicating if a slug test is used in the analysis.

Value

A list with the following entries:

method: A character string specifying the method used in the CI estimation (jackniffe)

hydraulic.parameters.names: A character vector with the names of the hydraulic parameters.

hydraulic.parameters.ci: A matrix with the requested confidence of the hydraulic parameters.

See also