cooper_WF_LT

Cooper's well function for slug tests in the Laplace domain

cooper_WF_LT(p, arg1, arg2)

Arguments

p
Laplace parameter

arg1
Value of the dimensionless wellbore storage coefficient (cd)

arg2
Value or the dimensionless radius (rd)

Value

The value of the Cooper's solution for slug test in the Laplace domain

References

Cooper, H. H.; Bredehoeft, J. D. & Papadopulos, I. S. Response of a finite-diameter well to an instantaneous charge of water Water Resources Research, 1967, 3, 263-269

See also

Examples

p <- seq(1,10,0.1)
res <- cooper_WF_LT(p, 0.1, 0.1)

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

print(res)

#>  [1] 0.06537548 0.06287310 0.06064277 0.05863673 0.05681843 0.05515929
#>  [7] 0.05363659 0.05223200 0.05093051 0.04971973 0.04858926 0.04753034
#> [13] 0.04653553 0.04559843 0.04471351 0.04387599 0.04308168 0.04232689
#> [19] 0.04160838 0.04092326 0.04026896 0.03964320 0.03904392 0.03846926
#> [25] 0.03791754 0.03738726 0.03687703 0.03638560 0.03591182 0.03545464
#> [31] 0.03501310 0.03458631 0.03417347 0.03377380 0.03338662 0.03301129
#> [37] 0.03264719 0.03229378 0.03195054 0.03161697 0.03129264 0.03097711
#> [43] 0.03067000 0.03037092 0.03007954 0.02979552 0.02951856 0.02924836
#> [49] 0.02898466 0.02872720 0.02847572 0.02823001 0.02798984 0.02775501
#> [55] 0.02752532 0.02730058 0.02708062 0.02686527 0.02665438 0.02644778
#> [61] 0.02624533 0.02604690 0.02585236 0.02566158 0.02547443 0.02529081
#> [67] 0.02511061 0.02493371 0.02476003 0.02458946 0.02442191 0.02425730
#> [73] 0.02409553 0.02393654 0.02378023 0.02362654 0.02347539 0.02332672
#> [79] 0.02318045 0.02303652 0.02289488 0.02275545 0.02261819 0.02248304
#> [85] 0.02234995 0.02221886 0.02208972 0.02196249 0.02183712 0.02171356
#> [91] 0.02159178