cooper_solution_dlogt

Function to calculate the derivative of the drawdown with respect to the log of time for the Cooper solution for slug tests

cooper_solution_dlogt(ptest, cd, t0, t)

Arguments

ptest
A pumping_test object

cd
Value of the dimensionless storage coefficient (cd)

t0
Intercept of the straight line fitted to the drawdown data using the Cooper-Jacob approach

t
A numeric vector with the time values

Value

A numeric vector with the values of the drawdown

References

Cooper, H. H.; Bredehoeft, J. D. & Papadopulos, I. S. Response of a finite-diameter well to an instantaneous charge of water Water Resources Research, 1967, 3, 263-269

See also

Examples

data(cooper_slug)
ptest <- pumping_test("Well1", Q = 0, r = 0, t = cooper_slug$t, s = cooper_slug$s)
sol0 <- cooper_solution_initial(ptest)
sol_dlogt <- cooper_solution_dlogt(ptest, sol0$cd, sol0$t0, ptest$t)
print(sol_dlogt)

#>  [1] 0.0002406712 0.0003225016 0.0003881108 0.0004442859 0.0004959946
#>  [6] 0.0005435412 0.0005886951 0.0006319847 0.0006916124 0.0007528979
#> [11] 0.0008083501 0.0008816399 0.0009530151 0.0010232637 0.0011083538
#> [16] 0.0012108978 0.0013130821 0.0014153704 0.0015184126 0.0016389639
#> [21] 0.0017808845 0.0019440572 0.0021152585 0.0022908375 0.0024959663
#> [26] 0.0027117267 0.0029393195 0.0032035965 0.0034875217 0.0037885784
#> [31] 0.0041426436 0.0045249577 0.0049384082 0.0053862776 0.0061884181
#> [36] 0.0070529194 0.0076402833 0.0084067369 0.0092896412 0.0101594627
#> [41] 0.0110995795 0.0121918947 0.0133322218 0.0147199518 0.0161554235
#> [46] 0.0177708021 0.0195450576 0.0214606339 0.0235982770 0.0260425565
#> [51] 0.0287562514 0.0317072400 0.0349459839 0.0386014815 0.0426349645
#> [56] 0.0470484679 0.0520059910 0.0574999292 0.0634469409 0.0699829249
#> [61] 0.0771951495 0.0849890557 0.0934736812 0.1027124930 0.1126706432
#> [66] 0.1233111914 0.1345909682 0.1464938643