hantush_jacob_solution

Function to calculate the drawdown of a Hantush-Jacob model

hantush_jacob_solution(ptest, a, t0, lambda, t)

Arguments

ptest
A pumping_test object

a
Slope of the straight line fitted to the drawdown data using the Cooper-Jacob approach

t0
Intercept of the straight line fitted to the drawdown data using the Cooper-Jacob approach

lambda
Leakage factor defined as $$\lambda = \sqrt{\frac{b^{'}T}{k_{h}^{'}}}$$ where \(b^{'}\) and \(k_{h}^{'}\) are the thickness and the vertical hydraulic conductivity of the aquitard, and \(T\) is the transmissivity of the underlying aquifer.

t
Numeric vector with the time values

Value

A numeric vector with the calculated drawdown

References

Hantush, M. S. Flow to wells in aquifers separated by a semipervious layer. Journal of Geophysical Research, 1967, 72, 1709.

See also

Examples

data(hantush_jacob)
ptest <- pumping_test('Well1', Q = 6.309e-3, r = 3.048, t=hantush_jacob$t, s = hantush_jacob$s)
ptest$additional_parameters$B <- 6.096
sol0 <- hantush_jacob_solution_initial(ptest)
sol <- hantush_jacob_solution(ptest,sol0$a, sol0$t0, sol0$lambda, ptest$t)
print(sol)

#>  [1]  3.517636  6.729367  8.286237  9.312751 10.500157 12.693940 13.947941
#>  [8] 15.465327 16.371900 17.261667 17.680387 18.114619 18.645370 18.958488
#> [15] 19.189455 19.480920 19.610387 19.696786 19.748767 19.796907 19.820848
#> [22] 19.827117 19.828188 19.828036 19.827632 19.826985 19.826654 19.826526
#> [29] 19.826501 19.826526 19.826551 19.826611 19.826662 19.826713 19.826746
#> [36] 19.826777 19.826806 19.826823 19.826849 19.826858 19.826866 19.826869
#> [43] 19.826871