hantush_jacob_solution_logdt

Function to calculate the derivative of drawdown with respect to the log of time

hantush_jacob_solution_dlogt(ptest, a, t0, lambda, t)

Arguments

ptest
A pumping_test object

a
Slope of the straight line fitted to the drawdown data using the Cooper-Jacob approach

t0
Intercept of the straight line fitted to the drawdown data using the Cooper-Jacob approach

lambda
Leakage factor defined as $$\lambda = \sqrt{\frac{b^{'}T}{k_{h}^{'}}}$$ where \(b^{'}\) and \(k_{h}^{'}\) are the thickness and the vertical hydraulic conductivity of the aquitard, and \(T\) is the transmissivity of the underlying aquifer.

t
Numeric vector with the time values

Value

This function returns a list with the x and y values of the derivative of drawdown with respect to log of time

References

Hantush, M. S. Flow to wells in aquifers separated by a semipervious layer. Journal of Geophysical Research, 1967, 72, 1709.

See also

Examples

data(hantush_jacob)
ptest <- pumping_test('Well1', Q = 6.309e-3, r = 3.048, t=hantush_jacob$t, s = hantush_jacob$s)
ptest$additional_parameters$B <- 6.096
sol0 <- hantush_jacob_solution_initial(ptest)
sol_dlogt <- hantush_jacob_solution_dlogt(ptest, sol0$a, sol0$t0, sol0$lambda, ptest$t)
print(sol_dlogt)

#>  [1] 2.759518e+00 3.106500e+00 3.172513e+00 3.179628e+00 3.104207e+00
#>  [6] 3.052661e+00 2.891429e+00 2.705374e+00 2.471738e+00 2.256058e+00
#> [11] 2.063788e+00 1.785664e+00 1.489434e+00 1.245617e+00 9.409885e-01
#> [16] 6.549476e-01 4.663972e-01 3.238129e-01 1.946724e-01 8.413695e-02
#> [21] 2.670301e-02 5.785952e-03 1.865726e-04 3.352343e-04 4.033625e-04
#> [26] 4.162084e-04 3.801071e-04 3.290892e-04 2.779765e-04 2.195395e-04
#> [31] 1.696408e-04 1.145053e-04 6.850489e-05 3.935675e-05 1.924593e-05
#> [36] 8.915035e-06