hantush_jacob_well_function

Well function of the Hantush-Jacob model

hantush_jacob_well_function(td, par)

Arguments

td
Dimensionless time

par
List with the beta parameter and the coefficients used in the Stehfest inversion approach. The beta parameter is defined as $$\beta = \frac{r}{4} \sqrt{\frac{k_{h}^{'}S_{s}^{'}}{TS}}$$ where \(T\) and \(S\) are the transmissivity and storage coefficient of the aquifer, and \(k_{h}^{'}\) and \(S_{s}^{'}\) are the vertical hydraulic conductivity and the Storage coefficient of the aquitard respectively.

Value

The value of the dimensionless drawdown of the Hantush-Jacob model

References

Hantush, M. S. Flow to wells in aquifers separated by a semipervious layer. Journal of Geophysical Research, 1967, 72, 1709.

See also

Examples

td <- logseq(-2, 4, 50)
par <- list( beta = 0.5)
res <- hantush_jacob_well_function(td, par)
print(res)

#>  [1] -1.114426e-08 -9.763013e-08 -5.870255e-07 -2.356990e-06 -5.679147e-06
#>  [6] -4.152360e-06  2.198458e-05  8.632269e-05  1.391449e-04  8.887939e-05
#> [11]  1.449473e-04  1.556976e-03  7.362647e-03  2.249568e-02  5.293492e-02
#> [16]  1.042228e-01  1.800529e-01  2.814786e-01  4.068666e-01  5.523766e-01
#> [21]  7.126360e-01  8.813664e-01  1.051857e+00  1.217311e+00  1.371151e+00
#> [26]  1.507410e+00  1.621291e+00  1.709899e+00  1.772945e+00  1.813033e+00
#> [31]  1.835122e+00  1.845188e+00  1.848633e+00  1.849241e+00  1.849046e+00
#> [36]  1.848827e+00  1.848742e+00  1.848744e+00  1.848775e+00  1.848804e+00
#> [41]  1.848823e+00  1.848833e+00  1.848837e+00  1.848839e+00  1.848839e+00
#> [46]  1.848839e+00  1.848839e+00  1.848839e+00  1.848838e+00  1.848838e+00