jacob_lohman_F_calculate_parameters

Calculate hydraulic parameters of the Jacob-Lohman Solution

jacob_lohman_F_calculate_parameters(ptest, par, hydraulic = TRUE)

Arguments

ptest
A pumping_test object

par
A list with the Slope (a) and the Intercept (t0) of the straight line fitted to the drawdown data using the Cooper-Jacob approach

hydraulic
Logical flag to indicate if hydraulic parameters are calculated. If False, the the statistcal parameter (a and t0) are calculated.

Value

A list with

Tr: Transmisivity.

Ss: The Storage coefficient

radius_influence

References

Jacob, C. E. & Lohman, S. W. Nonsteady Flow to a Well Of Constant Drawdown. American Geophysical Union, 1952, 33, 10.

See also

Examples

data(jacob_lohman)
s <- vector("numeric", length(jacob_lohman$t))
s[1:length(jacob_lohman$t)] <- 0
ptest <- pumping_test("Well1", Q = jacob_lohman$q, r = 0.84, t = jacob_lohman$t, s = s)

#> Warning: la condición tiene longitud > 1 y sólo el primer elemento será usado

ptest$additional_parameters <- list(rw = 0.084, s0 = 28.142)
sol0 <- jacob_lohman_F_solution_initial(ptest)
jl.par <- jacob_lohman_F_calculate_parameters(ptest, sol0)
print(jl.par)

#> $Tr
#> [1] 1.899174e-07
#> 
#> $Ss
#> [1] 3.626905e-05
#> 
#> $radius_influence
#> [1] 1.121037
#>