jacob_lohman_F_solution

Function to calculate the drawdown in a constant rate test using the Jacob-Lohman Solution

jacob_lohman_F_solution(ptest, a, t0, t)

Arguments

ptest
A pumping_test object

a
Slope of the straight line fitted to the drawdown data using the Cooper-Jacob approach

t0
Intercept of the straight line fitted to the drawdown data using the Cooper-Jacob approach

t
Numeric vector with the times at which measurements were taken

Value

A numeric vector with the values of the drawdown

References

Jacob, C. E. & Lohman, S. W. Nonsteady Flow to a Well Of Constant Drawdown. American Geophysical Union, 1952, 33, 10.

See also

Examples

data(jacob_lohman)
s <- vector("numeric", length(jacob_lohman$t))
s[1:length(jacob_lohman$t)] <- 0
ptest <- pumping_test("Well1", Q = jacob_lohman$q, r = 0.84, t = jacob_lohman$t, s = s)

#> Warning: la condición tiene longitud > 1 y sólo el primer elemento será usado

ptest$additional_parameters <- list(rw = 0.084, s0 = 28.142)
sol0 <- jacob_lohman_F_solution_initial(ptest)
sol <- jacob_lohman_F_solution(ptest, sol0$a, sol0$t0, ptest$t)
print(sol)

#>  [1] 1.000000e-12 1.000000e-12 1.000000e-12 1.000000e-12 1.000000e-12
#>  [6] 1.000000e-12 6.549962e-05 5.606949e-04 1.048647e-03 3.012478e-04
#> [11] 1.000000e-12 2.431470e-03 2.529024e-02 7.074195e-02 1.426487e-01
#> [16] 2.869214e-01 4.607646e-01 6.125247e-01 7.438434e-01