papadopulos_cooper_calculate_parameters

Function to calculate the hydraulic parameters of the Papadopulos-Cooper solution

papadopulos_cooper_calculate_parameters(ptest, par, hydraulic = TRUE)

Arguments

ptest
A pumping test object

par
A list with the values of a (slope) and t0 (Intercept) of the straight line fitted to the drawdown data using the Cooper-Jacob approach.

hydraulic
Logical flag to indicate if hydraulic parameters are calculated. If False, the the statistcal parameters (a and t0) are calculated.

Value

A list with the hydraulic parameters:

Tr: Transmissivity

Ss: Storage coefficient

cd: Wellbore Storage

rho: Dimensionless radius

References

Papadopulos, I. S. & Cooper, H. H. Drawdown in a well of large diameter. Water Resources Research, 1967, 3, 241-244.

See also

Examples

data(papadopulos_cooper)
ptest <- pumping_test("Well1", Q=0.0050472, r = 3.048, t = as.numeric(papadopulos_cooper$t),
                      s = papadopulos_cooper$s)
par <- list(rw = 0.6096, rc = 0.6096 )
ptest$additional_parameters <- par
sol0 <- papadopulos_cooper_solution_initial(ptest)
pc.par <- papadopulos_cooper_calculate_parameters(ptest, sol0)
print(pc.par)

#> $Tr
#> [1] 0.0008732865
#> 
#> $Ss
#> [1] 0.004969532
#> 
#> $cd
#> [1] 100.6131
#>