pumpingtest: A package for the analysis and evaluation of aquifer tests

The pumpingtest package provides functions to analyze and evaluate aquifer tests.

Details

The package includes the solution to the most common pumping tests applied in aquifers such as:

Theis and Cooper-Jacob solutions (confined aquifer)

Hantush-Jacob solution (leaky aquifer)

Jacob-Lohman solution (artesian confined aquifer)

Boulton solution (phreatic aquifer)

Cooper solution (slug tests)

Agarwal solution (recovery tests)

Agarwal solution (tests with skin effects)

General Radial Flow (fractured aquifer)

Neuzil solution (pulse tests)

Papadopoulous-Cooper solution (large diameter wells)

Warren and Root solution (dual porosity aquifer)

Gringarten solution (single fracture)

Hvorslev solution (slug tests)

Bower-Rice solution (slug tests)

The estimation of the hydraulic parameter is achieved by different optimization algorithms including

Nonlinear least squares with the Levenberg-Marquart algorithm (package minpack.lm)

L-BFGS-B (package optim)

Simulated Annealing (package GenSA)

Genetic Algorithms(package GA)

Particle Swarm Optimization (package PSO)

Differential Evolution (package DEoptim)

Estimation of Distribution Algorithm using Copulas (package copulaedas)

with four possible objective functions:

Sum of Squares Residuals

Mean Absolute Deviation

Maximum Absolute Deviation

Maximum Likelihood (under the assumption that the residuals follow a normal distribution).

Adaptative MCMC Auxiliary functions

prior, posterior, proposalfunction_cov, proposalfunction, run_adap_metropolis_MCMC

Agarwal functions (Recovery tests)

The functions included in this section are:

agarwal_recovery_well_function, agarwal_recovery_solution_initial, agarwal_recovery_calculate_parameters, agarwal_recovery_solution, agarwal_recovery_WF_LT, agarwal_recovery_solution_dlogt, equivalent_time

Agarwal functions (Pumping tests with skin effect)

The functions included in this section are:

agarwal_skin_well_function, agarwal_skin_solution_initial, agarwal_skin_calculate_parameters, agarwal_skin_solution, agarwal_skin_WF_LT, agarwal_skin_solution_dlogt

Aquifer functions

The aquifer functions include several functions designed to represent several types of aquifers including:

Boulton functions

boulton_well_function, boulton_solution_initial, boulton_calculate_parameters, boulton_solution, boulton_solution_dlogt, boulton_WF_LT

Bouwer Rice functions

These functions are used in the estimation of hydraulic parameters using a slug test

The functions included in this section are:

Acoeff, Bcoeff, Ccoeff, bouwer_rice_well_function, bouwer_rice_solution_initial, bouwer_rice_calculate_parameters, bouwer_rice_solution, bouwer_rice_solution_dlogt

Cooper-Jacob functions

cooper_jacob_well_function, cooper_jacob_solution_initial, cooper_jacob_calculate_parameters, cooper_jacob_solution

Cooper functions

cooper_well_function, cooper_solution_initial, cooper_calculate_parameters, cooper_solution, cooper_solution_dlogt, cooper_WF_LT

Equivalent_time functions

These functions are used in the estimation of the aquifer parameters for the recovery and variable rate tests.

The functions included in this section are: recovery_equivalent_time, variable_rate_equivalent_time

Flow Dimension functions

flow_dimension

General_radial_flow functions

The functions included in this section are:

general_radial_flow_well_function, general_radial_flow_solution_initial, general_radial_flow_calculate_parameters, general_radial_flow_solution, general_radial_flow_WF_LT, general_radial_flow_dlogt

Gringarten functions

These functions are used in the estimation of pumping tests in a single fracture aquifers.

The functions included in this section are:

gringarten_well_function, gringarten_solution_initial, gringarten_calculate_parameters, gringarten_solution, gringarten_solution_dlogt

Hantush-Jacob functions

hantush_jacob_well_function, hantush_jacob_solution_initial, hantush_jacob_calculate_parameters, hantush_jacob_solution, hantush_jacob_solution_dlogt, hantush_jacob_WF_LT

Hvorslev functions

These functions are used in the estimation of the aquifer parameters using a slug test.

The functions included in this section are:

hvorslev_well_function, hvorslev_solution_initial, hvorslev_calculate_parameters, hvorslev_solution, hvorslev_solution_dlogt

jacob_lohman functions

There are two types of functions: F and G. The Jacob_Lohman F functions are used to calculate drawdown at specific distance from the well, while the Jacob_Lohman G Functions are used to calculate the discharge at the well or a given distance from it.

The F functions are:

jacob_lohman_F_well_function, jacob_lohman_F_solution_initial, jacob_lohman_F_calculate_parameters, jacob_lohman_F_solution, jacob_lohman_F_solution_dlogt, jacob_lohman_F_WF_LT

The G Functions are

jacob_lohman_G_well_function, jacob_lohman_G_solution_initial, jacob_lohman_G_calculate_parameters, jacob_lohman_G_solution, jacob_lohman_G_solution_dlogt, jacob_lohman_G_WF_LT

Laplace Inversion functions

stehfest_coeffs, stehfest_inversion

Neuzil functions

These functions are used in the estimation of pulse tests.

The functions included in this section are: neuzil_well_function, neuzil_solution_initial, neuzil_calculate_parameters, neuzil_solution, neuzil_WF_LT, neuzil_solution_dlogt

Objective Function functions

residual_sum_squares, loglikelihood, bc_transf, loglikelihood_bc, mean_absolute_deviation, max_absolute_deviation, mean_square_error

Papadopoulous-Cooper functions

papadopulos_cooper_well_function, papadopulos_cooper_solution_initial, papadopulos_cooper_calculate_parameters, papadopulos_cooper_solution, papadopulos_cooper_solution_dlogt, papadopulos_cooper_WF_LT

Base functions

The base function includes the contructor of the S3 class pumping_test and associated functions to display summaries, print on the console, create diagnostic and estimation plots with the information from a pumping test, estimate the hydraulic parameters, and predict.

The functions in this section are:

pumping_test, summary, print, plot, fit, evaluate, simulate, confint

Smoothing functions

bandwidth_lpridge, smoothing.pumping_test

Theis functions

The Theis solution describes the drawdown variations in a confined, homogeneous, isotropic, aquifer of infinite horizontal extent due to the water extraction from a well that fully penetrates the aquifer.

The Functions included here are:

theis_well_function, theis_solution_initial, theis_calculate_parameters, theis_solution, theis_solution_dlogt, rtheis_solution

twalk Auxiliary functions

psi1, traverse_kernel, walk_kernel, hop_kernel, blow_kernel, log_ratio_density_blow, log_ratio_density_hop, run_twalk_MCMC

Utility functions

logseq, log_derivative, log_derivative_central, log_derivative_bourdet, log_derivative_spline, log_derivative_spane, log_derivative_smoothspline, log_derivative_kernel_reg, log_derivative_locpol, log_derivative_lokern, log_derivative_lpridge

Warren and Root functions

These functions are used in the estimation of pumping tests in dual porosity aquifers.

The functions included in this section are:

warren_root_well_function, warren_root_solution_initial, warren_root_calculate_parameters, warren_root_solution, warren_root_WF_LT, warren_root_solution_dlogt